requirements.txt
numpy
pandas
matplotlib
streamlit
scikit-learn
streamlit-drawable-canvas
seaborn
japanize-matplotlib
cupy-cuda12x
nvidia-cublas-cu12
nvidia-cusolver-cu12
nvidia-curand-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cuda-nvrtc-cu12
